(#32) design : 수동 결제 쿠폰함 UI
