--- src/components/coupon/ManualPaymentCoupon.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ManualPaymentCoupon = () => {
-  return <div>ManualPaymentCoupon</div>;
-};
-
-export default ManualPaymentCoupon;
